Move common parts of the protocol to dedicated file
